AST: Move a bunch of random @objc diagnostics to Sema
