+++ lib/Sema/TypeCheckDeclObjC.cpp
@@ -1736,3 +1736,630 @@ void markAsObjC(ValueDecl *D, ObjCReason reason,
     attr->setSwift3Inferred();
   }
 }
+
+void swift::diagnoseAttrsRequiringFoundation(SourceFile &SF) {
+  auto &Ctx = SF.getASTContext();
+
+  bool ImportsFoundationModule = false;
+
+  if (Ctx.LangOpts.EnableObjCInterop) {
+    if (!Ctx.LangOpts.EnableObjCAttrRequiresFoundation)
+      return;
+    if (SF.Kind == SourceFileKind::SIL)
+      return;
+  }
+
+  SF.forAllVisibleModules([&](ModuleDecl::ImportedModule import) {
+    if (import.second->getName() == Ctx.Id_Foundation)
+      ImportsFoundationModule = true;
+  });
+
+  if (ImportsFoundationModule)
+    return;
+
+  for (auto Attr : SF.AttrsRequiringFoundation) {
+    if (!Ctx.LangOpts.EnableObjCInterop)
+      Ctx.Diags.diagnose(Attr->getLocation(), diag::objc_interop_disabled)
+        .fixItRemove(Attr->getRangeWithAt());
+    Ctx.Diags.diagnose(Attr->getLocation(),
+                       diag::attr_used_without_required_module,
+                       Attr, Ctx.Id_Foundation)
+      .highlight(Attr->getRangeWithAt());
+  }
+}
+
+/// Compute the information used to describe an Objective-C redeclaration.
+std::pair<unsigned, DeclName> swift::getObjCMethodDiagInfo(
+                                AbstractFunctionDecl *member) {
+  if (isa<ConstructorDecl>(member))
+    return { 0 + member->isImplicit(), member->getFullName() };
+
+  if (isa<DestructorDecl>(member))
+    return { 2 + member->isImplicit(), member->getFullName() };
+
+  if (auto accessor = dyn_cast<AccessorDecl>(member)) {
+    switch (accessor->getAccessorKind()) {
+#define OBJC_ACCESSOR(ID, KEYWORD)
+#define ACCESSOR(ID) \
+    case AccessorKind::ID:
+#include "swift/AST/AccessorKinds.def"
+      llvm_unreachable("Not an Objective-C entry point");
+
+    case AccessorKind::Get:
+      if (auto var = dyn_cast<VarDecl>(accessor->getStorage()))
+        return { 5, var->getFullName() };
+
+      return { 6, Identifier() };
+
+    case AccessorKind::Set:
+      if (auto var = dyn_cast<VarDecl>(accessor->getStorage()))
+        return { 7, var->getFullName() };
+      return { 8, Identifier() };
+    }
+
+    llvm_unreachable("Unhandled AccessorKind in switch.");
+  }
+
+  // Normal method.
+  auto func = cast<FuncDecl>(member);
+  return { 4, func->getFullName() };
+}
+
+bool swift::fixDeclarationName(InFlightDiagnostic &diag, ValueDecl *decl,
+                               DeclName targetName) {
+  if (decl->isImplicit()) return false;
+  if (decl->getFullName() == targetName) return false;
+
+  // Handle properties directly.
+  if (auto var = dyn_cast<VarDecl>(decl)) {
+    // Replace the name.
+    SmallString<64> scratch;
+    diag.fixItReplace(var->getNameLoc(), targetName.getString(scratch));
+    return false;
+  }
+
+  // We only handle functions from here on.
+  auto func = dyn_cast<AbstractFunctionDecl>(decl);
+  if (!func) return true;
+
+  auto name = func->getFullName();
+
+  // Fix the name of the function itself.
+  if (name.getBaseName() != targetName.getBaseName()) {
+    diag.fixItReplace(func->getLoc(), targetName.getBaseName().userFacingName());
+  }
+
+  // Fix the argument names that need fixing.
+  assert(name.getArgumentNames().size()
+          == targetName.getArgumentNames().size());
+  auto params = func->getParameters();
+  for (unsigned i = 0, n = name.getArgumentNames().size(); i != n; ++i) {
+    auto origArg = name.getArgumentNames()[i];
+    auto targetArg = targetName.getArgumentNames()[i];
+
+    if (origArg == targetArg)
+      continue;
+
+    auto *param = params->get(i);
+
+    // The parameter has an explicitly-specified API name, and it's wrong.
+    if (param->getArgumentNameLoc() != param->getLoc() &&
+        param->getArgumentNameLoc().isValid()) {
+      // ... but the internal parameter name was right. Just zap the
+      // incorrect explicit specialization.
+      if (param->getName() == targetArg) {
+        diag.fixItRemoveChars(param->getArgumentNameLoc(),
+                              param->getLoc());
+        continue;
+      }
+
+      // Fix the API name.
+      StringRef targetArgStr = targetArg.empty()? "_" : targetArg.str();
+      diag.fixItReplace(param->getArgumentNameLoc(), targetArgStr);
+      continue;
+    }
+
+    // The parameter did not specify a separate API name. Insert one.
+    if (targetArg.empty())
+      diag.fixItInsert(param->getLoc(), "_ ");
+    else {
+      llvm::SmallString<8> targetArgStr;
+      targetArgStr += targetArg.str();
+      targetArgStr += ' ';
+      diag.fixItInsert(param->getLoc(), targetArgStr);
+    }
+  }
+
+  return false;
+}
+
+bool swift::fixDeclarationObjCName(InFlightDiagnostic &diag, ValueDecl *decl,
+                                   Optional<ObjCSelector> targetNameOpt,
+                                   bool ignoreImpliedName) {
+  if (decl->isImplicit())
+    return false;
+
+  // Subscripts cannot be renamed, so handle them directly.
+  if (isa<SubscriptDecl>(decl)) {
+    diag.fixItInsert(decl->getAttributeInsertionLoc(/*forModifier=*/false),
+                     "@objc ");
+    return false;
+  }
+
+  // Determine the Objective-C name of the declaration.
+  ObjCSelector name = *decl->getObjCRuntimeName();
+  auto targetName = *targetNameOpt;
+
+  // Dig out the existing '@objc' attribute on the witness. We don't care
+  // about implicit ones because they don't have useful source location
+  // information.
+  auto attr = decl->getAttrs().getAttribute<ObjCAttr>();
+  if (attr && attr->isImplicit())
+    attr = nullptr;
+
+  // If there is an @objc attribute with an explicit, incorrect witness
+  // name, go fix the witness name.
+  if (attr && name != targetName &&
+      attr->hasName() && !attr->isNameImplicit()) {
+    // Find the source range covering the full name.
+    SourceLoc startLoc;
+    if (attr->getNameLocs().empty())
+      startLoc = attr->getRParenLoc();
+    else
+      startLoc = attr->getNameLocs().front();
+
+    // Replace the name with the name of the requirement.
+    SmallString<64> scratch;
+    diag.fixItReplaceChars(startLoc, attr->getRParenLoc(),
+                           targetName.getString(scratch));
+    return false;
+  }
+
+  // We need to create or amend an @objc attribute with the appropriate name.
+
+  // Form the Fix-It text.
+  SourceLoc startLoc;
+  SmallString<64> fixItText;
+  {
+    assert((!attr || !attr->hasName() || attr->isNameImplicit() ||
+            name == targetName) && "Nothing to diagnose!");
+    llvm::raw_svector_ostream out(fixItText);
+
+    // If there is no @objc attribute, we need to add our own '@objc'.
+    if (!attr) {
+      startLoc = decl->getAttributeInsertionLoc(/*forModifier=*/false);
+      out << "@objc";
+    } else {
+      startLoc = Lexer::getLocForEndOfToken(decl->getASTContext().SourceMgr,
+                                            attr->getRange().End);
+    }
+
+    // If the names of the witness and requirement differ, we need to
+    // specify the name.
+    if (name != targetName || ignoreImpliedName) {
+      out << "(";
+      out << targetName;
+      out << ")";
+    }
+
+    if (!attr)
+      out << " ";
+  }
+
+  diag.fixItInsert(startLoc, fixItText);
+  return false;
+}
+
+namespace {
+  /// Produce a deterministic ordering of the given declarations.
+  class OrderDeclarations {
+    SourceManager &SrcMgr;
+
+  public:
+    OrderDeclarations(SourceManager &srcMgr) : SrcMgr(srcMgr) { }
+
+    bool operator()(ValueDecl *lhs, ValueDecl *rhs) const {
+      // If the declarations come from different modules, order based on the
+      // module.
+      ModuleDecl *lhsModule = lhs->getDeclContext()->getParentModule();
+      ModuleDecl *rhsModule = rhs->getDeclContext()->getParentModule();
+      if (lhsModule != rhsModule) {
+        return lhsModule->getName().str() < rhsModule->getName().str();
+      }
+
+      // If the two declarations are in the same source file, order based on
+      // location within that source file.
+      SourceFile *lhsSF = lhs->getDeclContext()->getParentSourceFile();
+      SourceFile *rhsSF = rhs->getDeclContext()->getParentSourceFile();
+      if (lhsSF == rhsSF) {
+        // If only one location is valid, the valid location comes first.
+        if (lhs->getLoc().isValid() != rhs->getLoc().isValid()) {
+          return lhs->getLoc().isValid();
+        }
+
+        // Prefer the declaration that comes first in the source file.
+        return SrcMgr.isBeforeInBuffer(lhs->getLoc(), rhs->getLoc());
+      }
+
+      // The declarations are in different source files (or unknown source
+      // files) of the same module. Order based on name.
+      // FIXME: This isn't a total ordering.
+      return lhs->getFullName() < rhs->getFullName();
+    }
+  };
+} // end anonymous namespace
+
+/// Lookup for an Objective-C method with the given selector in the
+/// given class or any of its superclasses.
+static AbstractFunctionDecl *lookupObjCMethodInClass(
+                               ClassDecl *classDecl,
+                               ObjCSelector selector,
+                               bool isInstanceMethod,
+                               bool isInitializer,
+                               SourceManager &srcMgr,
+                               bool inheritingInits = true) {
+  if (!classDecl)
+    return nullptr;
+
+  // Look for an Objective-C method in this class.
+  auto methods = classDecl->lookupDirect(selector, isInstanceMethod);
+  if (!methods.empty()) {
+    // If we aren't inheriting initializers, remove any initializers from the
+    // list.
+    if (!inheritingInits &&
+        std::find_if(methods.begin(), methods.end(),
+                     [](AbstractFunctionDecl *func) {
+                       return isa<ConstructorDecl>(func);
+                     }) != methods.end()) {
+      SmallVector<AbstractFunctionDecl *, 4> nonInitMethods;
+      std::copy_if(methods.begin(), methods.end(),
+                   std::back_inserter(nonInitMethods),
+                   [&](AbstractFunctionDecl *func) {
+                     return !isa<ConstructorDecl>(func);
+                   });
+      if (nonInitMethods.empty())
+        return nullptr;
+
+      return *std::min_element(nonInitMethods.begin(), nonInitMethods.end(),
+                               OrderDeclarations(srcMgr));
+    }
+
+    return *std::min_element(methods.begin(), methods.end(),
+                             OrderDeclarations(srcMgr));
+  }
+
+  // Recurse into the superclass.
+  if (!classDecl->hasSuperclass())
+    return nullptr;
+
+  // Determine whether we are (still) inheriting initializers.
+  inheritingInits = inheritingInits &&
+                    classDecl->inheritsSuperclassInitializers(nullptr);
+  if (isInitializer && !inheritingInits)
+    return nullptr;
+
+  return lookupObjCMethodInClass(classDecl->getSuperclassDecl(), selector,
+                                 isInstanceMethod, isInitializer, srcMgr,
+                                 inheritingInits);
+}
+
+bool swift::diagnoseUnintendedObjCMethodOverrides(SourceFile &sf) {
+  auto &Ctx = sf.getASTContext();
+  auto &methods = sf.ObjCMethodList;
+
+  // If no Objective-C methods were defined in this file, we're done.
+  if (methods.empty())
+    return false;
+
+  // Sort the methods by declaration order.
+  std::sort(methods.begin(), methods.end(), OrderDeclarations(Ctx.SourceMgr));
+
+  // For each Objective-C method declared in this file, check whether
+  // it overrides something in one of its superclasses. We
+  // intentionally don't respect access control here, since everything
+  // is visible to the Objective-C runtime.
+  bool diagnosedAny = false;
+  for (auto method : methods) {
+    // If the method has an @objc override, we don't need to do any
+    // more checking.
+    if (auto overridden = method->getOverriddenDecl()) {
+      if (overridden->isObjC())
+        continue;
+    }
+
+    // Skip deinitializers.
+    if (isa<DestructorDecl>(method))
+      continue;
+
+    // Skip invalid declarations.
+    if (method->isInvalid())
+      continue;
+
+    // Skip declarations with an invalid 'override' attribute on them.
+    if (auto attr = method->getAttrs().getAttribute<OverrideAttr>(true)) {
+      if (attr->isInvalid())
+        continue;
+    }
+
+    auto classDecl = method->getDeclContext()->getSelfClassDecl();
+    if (!classDecl)
+      continue; // error-recovery path, only
+
+    if (!classDecl->hasSuperclass())
+      continue;
+
+    // Look for a method that we have overridden in one of our
+    // superclasses.
+    // Note: This should be treated as a lookup for intra-module dependency
+    // purposes, but a subclass already depends on its superclasses and any
+    // extensions for many other reasons.
+    auto selector = method->getObjCSelector();
+    AbstractFunctionDecl *overriddenMethod
+      = lookupObjCMethodInClass(classDecl->getSuperclassDecl(),
+                                selector,
+                                method->isObjCInstanceMethod(),
+                                isa<ConstructorDecl>(method),
+                                Ctx.SourceMgr);
+    if (!overriddenMethod)
+      continue;
+
+    // Ignore stub implementations.
+    if (auto overriddenCtor = dyn_cast<ConstructorDecl>(overriddenMethod)) {
+      if (overriddenCtor->hasStubImplementation())
+        continue;
+    }
+
+    // Diagnose the override.
+    auto methodDiagInfo = getObjCMethodDiagInfo(method);
+    auto overriddenDiagInfo = getObjCMethodDiagInfo(overriddenMethod);
+    Ctx.Diags.diagnose(method, diag::objc_override_other,
+                       methodDiagInfo.first,
+                       methodDiagInfo.second,
+                       overriddenDiagInfo.first,
+                       overriddenDiagInfo.second,
+                       selector,
+                       overriddenMethod->getDeclContext()
+                         ->getDeclaredInterfaceType());
+    const ValueDecl *overriddenDecl = overriddenMethod;
+    if (overriddenMethod->isImplicit())
+      if (auto accessor = dyn_cast<AccessorDecl>(overriddenMethod))
+        overriddenDecl = accessor->getStorage();
+    Ctx.Diags.diagnose(overriddenDecl, diag::objc_declared_here,
+                       overriddenDiagInfo.first, overriddenDiagInfo.second);
+
+    diagnosedAny = true;
+  }
+
+  return diagnosedAny;
+}
+
+/// Retrieve the source file for the given Objective-C member conflict.
+static MutableArrayRef<AbstractFunctionDecl *>
+getObjCMethodConflictDecls(const SourceFile::ObjCMethodConflict &conflict) {
+  ClassDecl *classDecl = std::get<0>(conflict);
+  ObjCSelector selector = std::get<1>(conflict);
+  bool isInstanceMethod = std::get<2>(conflict);
+
+  return classDecl->lookupDirect(selector, isInstanceMethod);
+}
+
+/// Given a set of conflicting Objective-C methods, remove any methods
+/// that are legitimately overridden in Objective-C, i.e., because
+/// they occur in different modules, one is defined in the class, and
+/// the other is defined in an extension (category) thereof.
+static void removeValidObjCConflictingMethods(
+              MutableArrayRef<AbstractFunctionDecl *> &methods) {
+  // Erase any invalid or stub declarations. We don't want to complain about
+  // them, because we might already have complained about
+  // redeclarations based on Swift matching.
+  auto newEnd = std::remove_if(methods.begin(), methods.end(),
+                               [&](AbstractFunctionDecl *method) {
+                                 if (method->isInvalid())
+                                   return true;
+
+                                 if (auto ad = dyn_cast<AccessorDecl>(method)) {
+                                   return ad->getStorage()->isInvalid();
+                                 } 
+                                 
+                                 if (auto ctor 
+                                       = dyn_cast<ConstructorDecl>(method)) {
+                                   if (ctor->hasStubImplementation())
+                                     return true;
+
+                                   return false;
+                                 }
+
+                                 return false;
+                               });
+  methods = methods.slice(0, newEnd - methods.begin());
+}
+
+bool swift::diagnoseObjCMethodConflicts(SourceFile &sf) {
+  // If there were no conflicts, we're done.
+  if (sf.ObjCMethodConflicts.empty())
+    return false;
+
+  auto &Ctx = sf.getASTContext();
+
+  OrderDeclarations ordering(Ctx.SourceMgr);
+
+  // Sort the set of conflicts so we get a deterministic order for
+  // diagnostics. We use the first conflicting declaration in each set to
+  // perform the sort.
+  auto localConflicts = sf.ObjCMethodConflicts;
+  std::sort(localConflicts.begin(), localConflicts.end(),
+            [&](const SourceFile::ObjCMethodConflict &lhs,
+                const SourceFile::ObjCMethodConflict &rhs) {
+              return ordering(getObjCMethodConflictDecls(lhs)[1],
+                              getObjCMethodConflictDecls(rhs)[1]);
+            });
+
+  // Diagnose each conflict.
+  bool anyConflicts = false;
+  for (const auto &conflict : localConflicts) {
+    ObjCSelector selector = std::get<1>(conflict);
+
+    auto methods = getObjCMethodConflictDecls(conflict);
+
+    // Prune out cases where it is acceptable to have a conflict.
+    removeValidObjCConflictingMethods(methods);
+    if (methods.size() < 2)
+      continue;
+
+    // Diagnose the conflict.
+    anyConflicts = true;
+
+    // If the first method is in an extension and the second is not, swap them
+    // so the primary diagnostic is on the extension method.
+    if (isa<ExtensionDecl>(methods[0]->getDeclContext()) &&
+        !isa<ExtensionDecl>(methods[1]->getDeclContext())) {
+      std::swap(methods[0], methods[1]);
+
+    // Within a source file, use our canonical ordering.
+    } else if (methods[0]->getParentSourceFile() ==
+               methods[1]->getParentSourceFile() &&
+              !ordering(methods[0], methods[1])) {
+      std::swap(methods[0], methods[1]);
+    }
+
+    // Otherwise, fall back to the order in which the declarations were type
+    // checked.
+
+    auto originalMethod = methods.front();
+    auto conflictingMethods = methods.slice(1);
+
+    auto origDiagInfo = getObjCMethodDiagInfo(originalMethod);
+    for (auto conflictingDecl : conflictingMethods) {
+      auto diagInfo = getObjCMethodDiagInfo(conflictingDecl);
+
+      const ValueDecl *originalDecl = originalMethod;
+      if (originalMethod->isImplicit())
+        if (auto accessor = dyn_cast<AccessorDecl>(originalMethod))
+          originalDecl = accessor->getStorage();
+
+      if (diagInfo == origDiagInfo) {
+        Ctx.Diags.diagnose(conflictingDecl, diag::objc_redecl_same,
+                           diagInfo.first, diagInfo.second, selector);
+        Ctx.Diags.diagnose(originalDecl, diag::invalid_redecl_prev,
+                           originalDecl->getBaseName());
+      } else {
+        Ctx.Diags.diagnose(conflictingDecl, diag::objc_redecl,
+                           diagInfo.first,
+                            diagInfo.second,
+                           origDiagInfo.first,
+                           origDiagInfo.second,
+                           selector);
+        Ctx.Diags.diagnose(originalDecl, diag::objc_declared_here,
+                           origDiagInfo.first, origDiagInfo.second);
+      }
+    }
+  }
+
+  return anyConflicts;
+}
+
+/// Retrieve the source location associated with this declaration
+/// context.
+static SourceLoc getDeclContextLoc(DeclContext *dc) {
+  if (auto ext = dyn_cast<ExtensionDecl>(dc))
+    return ext->getLoc();
+
+  return cast<NominalTypeDecl>(dc)->getLoc();
+}
+
+bool swift::diagnoseObjCUnsatisfiedOptReqConflicts(SourceFile &sf) {
+  // If there are no unsatisfied, optional @objc requirements, we're done.
+  if (sf.ObjCUnsatisfiedOptReqs.empty())
+    return false;
+
+  auto &Ctx = sf.getASTContext();
+
+  // Sort the set of local unsatisfied requirements, so we get a
+  // deterministic order for diagnostics.
+  auto &localReqs = sf.ObjCUnsatisfiedOptReqs;
+  std::sort(localReqs.begin(), localReqs.end(),
+            [&](const SourceFile::ObjCUnsatisfiedOptReq &lhs,
+                const SourceFile::ObjCUnsatisfiedOptReq &rhs) -> bool {
+              return Ctx.SourceMgr.isBeforeInBuffer(getDeclContextLoc(lhs.first),
+                                                    getDeclContextLoc(rhs.first));
+            });
+
+  // Check each of the unsatisfied optional requirements.
+  bool anyDiagnosed = false;
+  for (const auto &unsatisfied : localReqs) {
+    // Check whether there is a conflict here.
+    ClassDecl *classDecl = unsatisfied.first->getSelfClassDecl();
+    auto req = unsatisfied.second;
+    auto selector = req->getObjCSelector();
+    bool isInstanceMethod = req->isInstanceMember();
+    // FIXME: Also look in superclasses?
+    auto conflicts = classDecl->lookupDirect(selector, isInstanceMethod);
+    if (conflicts.empty())
+      continue;
+
+    // Diagnose the conflict.
+    auto reqDiagInfo = getObjCMethodDiagInfo(unsatisfied.second);
+    auto conflictDiagInfo = getObjCMethodDiagInfo(conflicts[0]);
+    auto protocolName
+      = cast<ProtocolDecl>(req->getDeclContext())->getFullName();
+    Ctx.Diags.diagnose(conflicts[0],
+                       diag::objc_optional_requirement_conflict,
+                       conflictDiagInfo.first,
+                       conflictDiagInfo.second,
+                       reqDiagInfo.first,
+                       reqDiagInfo.second,
+                       selector,
+                       protocolName);
+
+    // Fix the name of the witness, if we can.
+    if (req->getFullName() != conflicts[0]->getFullName() &&
+        req->getKind() == conflicts[0]->getKind() &&
+        isa<AccessorDecl>(req) == isa<AccessorDecl>(conflicts[0])) {
+      // They're of the same kind: fix the name.
+      unsigned kind;
+      if (isa<ConstructorDecl>(req))
+        kind = 1;
+      else if (auto accessor = dyn_cast<AccessorDecl>(req))
+        kind = isa<SubscriptDecl>(accessor->getStorage()) ? 3 : 2;
+      else if (isa<FuncDecl>(req))
+        kind = 0;
+      else {
+        llvm_unreachable("unhandled @objc declaration kind");
+      }
+
+      auto diag = Ctx.Diags.diagnose(conflicts[0],
+                                     diag::objc_optional_requirement_swift_rename,
+                                     kind, req->getFullName());
+
+      // Fix the Swift name.
+      fixDeclarationName(diag, conflicts[0], req->getFullName());
+
+      // Fix the '@objc' attribute, if needed.
+      if (!conflicts[0]->canInferObjCFromRequirement(req))
+        fixDeclarationObjCName(diag, conflicts[0], req->getObjCRuntimeName(),
+                               /*ignoreImpliedName=*/true);
+    }
+
+    // @nonobjc will silence this warning.
+    bool hasExplicitObjCAttribute = false;
+    if (auto objcAttr = conflicts[0]->getAttrs().getAttribute<ObjCAttr>())
+      hasExplicitObjCAttribute = !objcAttr->isImplicit();
+    if (!hasExplicitObjCAttribute)
+      Ctx.Diags.diagnose(conflicts[0], diag::req_near_match_nonobjc, true)
+        .fixItInsert(
+          conflicts[0]->getAttributeInsertionLoc(/*forModifier=*/false),
+          "@nonobjc ");
+
+    Ctx.Diags.diagnose(getDeclContextLoc(unsatisfied.first),
+                       diag::protocol_conformance_here,
+                       true,
+                       classDecl->getFullName(),
+                      protocolName);
+    Ctx.Diags.diagnose(req, diag::kind_declname_declared_here,
+                       DescriptiveDeclKind::Requirement, reqDiagInfo.second);
+
+    anyDiagnosed = true;
+  }
+
+  return anyDiagnosed;
+}
